+++ EngineKit/Sources/DatabaseService/Interface/Protocols/ContainerManagerType.swift
@@ -9,9 +9,9 @@ public protocol ContainerManagerType {
     typealias Entry = ContainerModel
     
     @discardableResult
-    func createEntry(of size: Double) async throws -> Entry
+    func createEntry(of size: Int) async throws -> Entry
     
-    func update(_ entry: Entry, newSize: Double) async throws -> Entry
+    func update(_ entry: Entry, newSize: Int) async throws -> Entry
     func delete(_ entry: Entry) async throws
     
     func fetchAll() async throws -> [Entry]
